requirements.txt
facebook-business>=17.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
plotly>=5.20.0
pandas>=2.0.0
apify-client>=1.7.0
requests>=2.31.0
